requirements.txt
aiohttp==3.6.2     
astroid==2.4.2     
async-timeout==3.0.1     
attrs==19.3.0    
autopep8==1.5.3     
beautifulsoup4==4.9.1     
bs4==0.0.1     
certifi==2020.6.20 
chardet==3.0.4     
discord.py==1.3.3     
idna==2.9       
iso8601==0.1.11    
isort==4.3.21    
lazy-object-proxy==1.4.3     
mccabe==0.6.1     
multidict==4.7.6     
pip==20.0.2    
pkg-resources==0.0.0     
pychallonge==1.0
pycodestyle==2.6.0     
pylint==2.5.3     
requests==2.24.0    
setuptools==44.0.0    
six==1.15.0    
soupsieve==2.0.1     
toml==0.10.1    
urllib3==1.25.9    
websockets==8.1       
wrapt==1.12.1    
yarl==1.4.2   